+++ src/runtime/option-preprocess/index.ts
@@ -1,10 +1,8 @@
-import { get } from '@antv/util';
-import { G2View, G2ViewTree } from '../types/options';
-import { convertScale } from './scale';
+import { G2ViewTree } from '../types/options';
+import { flow } from '../../utils/flow';
+import { columnWidthRatio } from './style';
 
-export function optionPreprocess<T extends G2ViewTree = G2ViewTree>(
-  options: T,
-): T {
+export function optionPreprocess(options: G2ViewTree): G2ViewTree {
   const convertedOptions = preprocess(options);
   // If there are children, recursively convert each child node.
   if (convertedOptions.children && Array.isArray(convertedOptions.children)) {
@@ -17,19 +15,7 @@ export function optionPreprocess<T extends G2ViewTree = G2ViewTree>(
 }
 
 // Entry point for all syntactic sugar functions.
-export function preprocess<T extends G2ViewTree = G2ViewTree>(options: T): T {
-  const { style, scale, type } = options;
-  const scaleOption: G2View = {};
-  // style: { columnWidthRatio: 0.2 } => scale: { x: { padding: 0.8 } }
-  const columnWidthRatio = get(style, 'columnWidthRatio');
-  if (columnWidthRatio && type === 'interval') {
-    scaleOption.x = {
-      ...scale?.x,
-      ...convertScale(columnWidthRatio),
-    };
-  }
-  return {
-    ...options,
-    scale: { ...scale, ...scaleOption },
-  };
+export function preprocess(options: G2ViewTree): G2ViewTree {
+  //@todo define a type for params of flow
+  return flow(columnWidthRatio)(options);
 }
