--- src/runtime/option-preprocess/scale.ts
@@ -1,5 +1,0 @@
-export function convertScale(columnWidthRatio: number) {
-  return {
-    padding: 1 - columnWidthRatio,
-  };
-}
